backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.4
Pillow>=10.0.0
celery>=5.3.0
redis>=4.5.0
channels>=4.0.0
channels-redis>=4.1.0
django-extensions>=3.2.0
